Makefile cleanup.
This change subsitutes arcane symbols like $@ and $< for English words, with the expected result that the new file behaves exactly like the old one.

Well, the point is to eliminate *redundant* English words.  The rules become shorter.  In cases where the same word appears among a target and/or prerequisite and/or recipe, changing it in one place changes it for all of them.

In addition to the refactoring, some comments are added, and some 'clean'
targets are defined to permit comprehensive or targeted removal of
generated output.