Multiple Makefile changes.
This version is altered to improve the accuracy of make's determination of whether a target should be rebuilt (notice the '|' designator for order-only prerequisites in a few places, which may need further fine-tuning).

Furthermore, various 'mkdir' operations are now consolidated using dependency chains, rather than recipes that attempt redundant directory creation.  Besides cleaning up the output, this should reduce any possibility of directory timestamps being tweaked (with consequent attempts to needlessly rebuild targets) for no good reason.

This version is also enhanced to build the reference genome, samfiles, and bamfiles, using "make bamfiles".  After additional testing is completed, the expectation is that the 'all' target will refer to 'bamfiles' rather than reference_genome.

Note that it may seem tempting, based on CPU capacity, to "make -j 2 bamfiles".  However bear in mind that this will start 2 parallel STAR processes that *each* want to consume 30G+ of RAM.  That may theoretically work on a workstation with 64G of RAM, but it cripples one that has only 48G.

An initial .PHONY target is added, but probably will need to be tweaked subsequently.